Moved trajectory to DriveSubsystem and mapped to B
