+++ src/main/kotlin/frc/robot/Controls.kt
@@ -3,6 +3,8 @@ package frc.robot
 import edu.wpi.first.wpilibj.XboxController
 import edu.wpi.first.wpilibj2.command.Command
 import edu.wpi.first.wpilibj2.command.InstantCommand
+import frc.robot.subsystems.drive.DriveSubsystem
+import frc.robot.subsystems.drive.TestTrajectory
 import org.ghrobotics.lib.commands.sequential
 import org.ghrobotics.lib.wrappers.hid.*
 
@@ -12,6 +14,7 @@ object Controls {
     val driverFalconXbox = driverControllerLowLevel.mapControls {
 
         button(kX).changeOn(InstantCommand(Runnable { "Hullo!" }))
+        button(kB).change(DriveSubsystem.followTrajectory(TestTrajectory.trajectory))
     }
 
     val operatorWPIJoystick = XboxController(1)
