+++ src/main/kotlin/frc/robot/subsystems/drive/TestTrajectory.kt
@@ -1,4 +1,4 @@
-package frc.robot.autonomous
+package frc.robot.subsystems.drive
 
 import edu.wpi.first.wpilibj.geometry.Pose2d
 import org.ghrobotics.lib.mathematics.twodim.geometry.Pose2d
